Introduces translations between monads.
Not a lowest-level base functor (UserF, MatchmakingF) can be interpreted
into STM, and the STM translated to IO when necessary.

The idea is that this way I can get transactional semantics at the exact
desired levels out of my handlers, and then just call translate at the
end and everything winds up in IO.

And for handlers that go straight to IO, of course there's a Translates
IO IO instance.