Switched Login to use email instead of username.
This also removes the unpleasant dependency of login on user.

Unfortunately, now user needs to depend on login, to create a login
record on user registration.  I don't see how it can be avoided.

This change doesn't include the needed changes to the user handler to
create login records.

I might data types a la carte it.